restructure code to move get_hsml higher up (for reuse in volume scatter functions)
