+++ swiftsimio/visualisation/__init__.py
@@ -5,4 +5,4 @@ Visualisation sub-module for swiftismio.
 from .projection import scatter, project_gas, project_gas_pixel_grid
 from .slice import slice_scatter as slice
 from .slice import slice_gas, slice_gas_pixel_grid
-from .smoothing_length_generation import generate_smoothing_lengths
+from .smoothing_length import generate_smoothing_lengths
