+++ swiftsimio/visualisation/slice_backends/nearest_neighbors.py
@@ -1,8 +1,6 @@
-from numpy import float64, float32, ndarray, linspace, array, stack, meshgrid, cbrt
+from numpy import float64, float32, ndarray, linspace, array, stack, meshgrid
 
-from swiftsimio import SWIFTDataset, cosmo_array
 from swiftsimio.optional_packages import KDTree, TREE_AVAILABLE
-from swiftsimio.visualisation.slice_backends.sph import get_hsml as sph_get_hsml
 
 
 def build_tree(
@@ -53,34 +51,6 @@ def build_tree(
     else:
         data = stack((x, y, z), axis=1)
         return KDTree(data)
-
-
-def get_hsml(data: SWIFTDataset) -> cosmo_array:
-    """
-    Computes a "smoothing length" as the 3rd root of the volume of the particles.
-    This scheme uses volume weighting when computing slices.
-
-    Parameters
-    ----------
-    data : SWIFTDataset
-        The Dataset from which slice will be extracted
-
-    Returns
-    -------
-    The extracted "smoothing lengths".
-    """
-    try:
-        hsml = cbrt(data.gas.volume)
-    except AttributeError:
-        try:
-            # Try computing the volumes explicitly?
-            masses = data.gas.masses
-            densities = data.gas.densities
-            hsml = cbrt(masses / densities)
-        except AttributeError:
-            # Fall back to SPH behavior if above didn't work...
-            hsml = sph_get_hsml(data)
-    return hsml
 
 
 def slice_scatter(
